Resolves the rust.yml conflict. The branch was opened 2026-06-24 and had
drifted 13 commits behind main.

Resolution keeps main's workflow in full and adds ONLY this PR's coverage
steps, so the net change against main is +10 lines and nothing is removed.
Deliberately NOT taken from the branch: deletion of the Cloud Run deploy
job (removing a deployment pipeline should be an explicit decision, not a
merge side effect) and removal of explanatory comments including the
RUSTSEC-2023-0071 risk-acceptance rationale, which documents a security
decision and must not be silently dropped.

Preserved from main: paths-ignore docs filter (#99), actions/checkout@v7
(#86), the deploy job, and all section comments.

+++ projects-service/src/lib/models.rs
@@ -123,7 +123,7 @@ pub struct ProjectLink {
     pub link_type: String,
     pub label: String,
     pub url: String,
-    pub sort_order: i64,
+    pub sort_order: i32,
     pub created_at: String,
     pub updated_at: String,
 }
